+++ CHANGELOG.rst
@@ -54,6 +54,9 @@ Changes:
 - Added a new thread-local API that allows binding values to a thread-local context explicitly without affecting the default behavior of ``bind()``.
   `#222 <https://github.com/hynek/structlog/issues/222>`_,
   `#225 <https://github.com/hynek/structlog/issues/225>`_,
+- Added ``pass_foreign_args`` argument to ``structlog.stdlib.ProcessorFormatter``.
+  It allows to pass a foreign log record's ``args`` attribute to the event dictionary under the ``positional_args`` key.
+  `#228 <https://github.com/hynek/structlog/issues/228>`_,
 
 
 ----